Move signature near backing class
